Charon.Backend.Fdo.DirectorySizes

Description

Provides functionality for handling the directorysizes file in the FDO spec.

Types

newtype DirectorySizes

Represents the directorysizes contents.

Constructors

MkDirectorySizes

Fields

unDirectorySizes :: Seq DirectorySizesEntry

data DirectorySizesEntry

directorysizes entry.

Constructors

MkDirectorySizesEntry

Fields

size :: Bytes B Natural

Directory size in bytes. This does not include the size of the directory itself.

time :: Natural

The time this directory was deleted. The units are milliseconds since the unix epoch.

fileName :: ByteString

Percent encoded filename.

Adding

appendEntry :: forall m env. (HasCallStack, HasTrashHome env, MonadCatch m, MonadLoggerNS m, MonadReader env m, MonadFileReader m, MonadFileWriter m, MonadPathReader m, MonadPathWriter m, MonadTime m) => DirectorySizesEntry -> m ()

Appends an entry to directorysizes.

appendEntryTrashHome :: forall m. (HasCallStack, MonadCatch m, MonadLoggerNS m, MonadFileReader m, MonadFileWriter m, MonadPathReader m, MonadPathWriter m, MonadTime m) => PathI TrashHome -> DirectorySizesEntry -> m ()

Appends an entry to directorysizes.

writeDirectorySizes :: (HasCallStack, HasTrashHome env, MonadCatch m, MonadFileWriter m, MonadLoggerNS m, MonadPathReader m, MonadPathWriter m, MonadReader env m, MonadTime m) => DirectorySizes -> m ()

Writes entries to directorysizes.

writeDirectorySizesTrashHome :: (HasCallStack, MonadCatch m, MonadFileWriter m, MonadLoggerNS m, MonadPathReader m, MonadPathWriter m, MonadTime m) => PathI TrashHome -> DirectorySizes -> m ()

Writes entries to directorysizes.

Removing

removeEntry :: (HasCallStack, HasTrashHome env, MonadCatch m, MonadReader env m, MonadFileReader m, MonadFileWriter m, MonadLoggerNS m, MonadPathReader m, MonadPathWriter m, MonadTime m) => ByteString -> m ()

Removes an entry from directory sizes.

Reading

readDirectorySizes :: (HasCallStack, HasTrashHome env, MonadFileReader m, MonadPathReader m, MonadReader env m, MonadThrow m) => m DirectorySizes

Reads directorysizes.

readDirectorySizesTrashHome :: (HasCallStack, MonadFileReader m, MonadPathReader m, MonadThrow m) => PathI TrashHome -> m DirectorySizes

Reads directorysizes.

getDirectorySizesPath :: (HasTrashHome env, MonadReader env m) => m OsPath