Charon.Backend.Json.PathData

Description

Provides PathData for use with the Json backend.

PathData

data PathData

Data for an Fdo path. Maintains an invariant that the original path is not the root nor is it empty.

Constructors

UnsafePathData

Fields

pathType :: PathTypeW

The path type.

fileName :: PathI TrashEntryFileName

The path to be used in the trash directory.

originalPath :: PathI TrashEntryOriginalPath

The original path on the file system.

created :: Timestamp

Time this entry was created.

size :: Bytes B Natural

The size.

toPathData :: (HasCallStack, MonadAsync m, MonadCatch m, MonadLoggerNS m, MonadPathReader m, MonadPosixCompat m, MonadTerminal m) => Timestamp -> PathI TrashHome -> PathI TrashEntryOriginalPath -> m (PathData, PathTypeW)

For a given filepath, attempts to capture the following data:

Canonical path.

Unique name to be used in the trash directory.

File/directory type.

toCorePathData :: PathData -> PathData

fromCorePathData :: PathData -> PathData