Charon.Data.Metadata

Description

Provides metadata functionality.

Documentation

data Metadata

Holds trash metadata.

Constructors

MkMetadata

Fields

numEntries :: Natural

Number of top level entries in the trash index. This should be the same as the index length.

numFiles :: Natural

Number of total files in the trash.

logSize :: SomeSize Double

Size of the log file.

size :: SomeSize Double

Total size of the trash directory.

empty :: Metadata

Empty metadata.