Charon.Data.PathData.Formatting

Description

Formats the default PathData.

Types

data PathDataFormat

Determines how to format a textual PathData.

Constructors

FormatMultiline
Formats each file on its own line.

FormatTabular Coloring (Maybe ColFormat) (Maybe ColFormat)
Formats all fields on the same line.

FormatSingleline Coloring
Formats each entry on a single line, no table.

_FormatMultiline :: Prism' PathDataFormat ()

_FormatTabular :: Prism' PathDataFormat (Coloring, Maybe ColFormat, Maybe ColFormat)

_FormatSingleline :: Prism' PathDataFormat Coloring

data ColFormat

Constructors

ColFormatFixed Natural
Fixed length format.

ColFormatMax
Format the column according to its longest entry, if possible.

_ColFormatFixed :: Prism' ColFormat Natural

_ColFormatMax :: Prism' ColFormat ()

data Coloring

Whether we should color list command.

Constructors

ColoringOn
Coloring on

ColoringOff
Coloring off

ColoringDetect
Attempt to detect if the terminal support colors.

Sort

data Sort

How to sort the index list.

Constructors

Name
Sort by name.

Size
Sort by size.

readSort :: MonadFail m => Text -> m Sort

sortFn :: Bool -> Sort -> PathData -> PathData -> Ordering

Format functions

formatTabularHeader :: Natural -> Natural -> Text

formatTabularHeaderColor :: Color -> Natural -> Natural -> Text

formatTabularRow :: Natural -> Natural -> PathData -> Text

formatTabularRowColor :: Color -> Natural -> Natural -> PathData -> Text

formatMultiline :: PathData -> Text

formatSingleline :: PathData -> Text

formatSinglelineColor :: Color -> PathData -> Text

Sorting

sortNameCreated :: PathData -> PathData -> Ordering

sortSizeName :: PathData -> PathData -> Ordering

sortReverse :: (a -> b -> Ordering) -> a -> b -> Ordering

Field lengths

formatFileNameLenMin :: Natural

formatOriginalPathLenMin :: Natural

minTableWidth :: Natural

Minimum length needed to display the table.

reservedLineLen :: Natural

For tabular formatting, this is the necessary width for the fixed columns:

type: 9 size: 7 created: 19 separators: 12

NOTE: The separators includes the adjacent whitespace i.e. one separator ' | ' counts for 3, and since we have 4 that makes 12.

This does not include the minimum necessary total space (i.e. minimum 4 for name and 8 for original).