Charon.Data.PathType

Description

Provides types.

Documentation

newtype PathTypeW

Wrapper for PathType, so that we can give a Hashable instance.

Constructors

MkPathTypeW

Fields

unPathTypeW :: PathType

deleteFn :: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPathWriter m) => PathTypeW -> OsPath -> m ()

existsFn :: (HasCallStack, MonadCatch m, MonadPathReader m) => PathTypeW -> OsPath -> m Bool

This function tests both existence and that the the path is of the specified type.

renameFn :: (HasCallStack, MonadPathWriter m) => PathTypeW -> OsPath -> OsPath -> m ()

copyPath

Arguments

:: (HasCallStack, MonadIORef m, MonadMask m, MonadPathReader m, MonadPathWriter m)

=> PathTypeW

-> OsPath
Old path to copy.

-> OsPath
Fully qualified name for file copy.

-> OsPath
The directory in which to copy, for a dir copy.

-> m ()