Charon.Data.Paths

Description

Provides functionality for distinguishing path types.

Types

newtype PathI (i :: PathIndex)

Indexed OsPath so that we can prevent mixing up different filepaths.

Constructors

MkPathI

Fields

unPathI :: OsPath

data PathIndex

Types of filepaths used in Charon.

Constructors

TrashHome
The trash directory.

TrashLog
The trash log file.

TrashDirFiles
The directory to the trash files themselves i.e. trash/files.

TrashDirInfo
The directory to the trash info files i.e. trash/info.

TrashEntryPath
The full trash path i.e. <trash-home>/files/'<trash-name>.

TrashEntryInfo
The full trash info path i.e. <trash-home>/info/'<trash-name>.trashinfo.

TrashEntryFileName
The name corresponding to some file/directory in the trash directory.

TrashEntryOriginalPath
The original path for some file/directory in the trash directory.

Functions

Specific

isEmpty :: MonadThrow m => PathI i -> m Bool

isRoot :: MonadThrow m => PathI i -> m Bool

Returns true if the path is the root.

isRoot' :: MonadThrow m => OsPath -> m Bool

Returns true if the path is the root.

isDots :: MonadThrow m => PathI i -> m Bool

Returns true if the path ends in nothing but dots e.g. pathtodots...

toString :: PathI i -> String

PathI to String. Attempts decoding for nicer display.

toText :: PathI i -> Text

PathI to Text via toString.

General

These functions allows for lifting arbitrary OsPath functions onto our PathI. Note that this can easily invalidate any invariants we would like to hold (e.g. appending a path can turn a directory into a file), so caution must be exercised.

showPaths :: [PathI a] -> String

Pretty-print a list of paths.

>>> showPaths ["one", "two"]
"one, two"

reindex :: PathI i -> PathI j

Modifies the index.

(<//>) :: PathI i1 -> PathI i2 -> PathI i3 infixr 5

(</>) lifted to PathI. Notice the index can change, so take care.

(<//) :: PathI i1 -> OsPath -> PathI i2 infixl 5

(//>) :: OsPath -> PathI i1 -> PathI i2 infixl 5

applyPathI :: HasCallStack => (HasCallStack => OsPath -> a) -> PathI i -> a

Lifts an OsPath function to PathI.

liftPathI :: (OsPath -> OsPath) -> PathI i -> PathI j

Lifts an OsPath transformation to PathI, allowing for the index to change.

liftPathI' :: (OsPath -> OsPath) -> PathI i -> PathI i

liftPathI specialized to the same index. This should be preferred as the former is easier to use incorrectly.

liftPathIF :: (Functor f, HasCallStack) => (HasCallStack => OsPath -> f OsPath) -> PathI i -> f (PathI j)

Lifts an effectful FilePath transformation to PathI.

liftPathIF' :: Functor f => (HasCallStack => OsPath -> f OsPath) -> PathI i -> f (PathI i)

liftPathIF specialized to the same index. This should be preferred as the former is easier to use incorrectly.