Charon.Exception

Description

Provides exceptions used by Charon.

Trash

Entries

General

newtype TrashEntryNotFoundE

Generic trash path not found error. Distinct from TrashEntryFileNotFoundE in that the latter indicates that the entry exists in trash/info but not trash/files, whereas this exception is less specific i.e. we found nothing in trash/info but did not look in trash/files.

Constructors

MkTrashEntryNotFoundE (PathI TrashEntryFileName)

newtype TrashEntryWildcardNotFoundE

Error for not finding any files via wildcard search.

Constructors

MkTrashEntryWildcardNotFoundE (PathI TrashEntryFileName)

Partial success

data TrashEntryFileNotFoundE

Path found in the index but not files.

Constructors

MkTrashEntryFileNotFoundE (PathI TrashHome) (PathI TrashEntryFileName)

data TrashEntryInfoNotFoundE

Path found files but not index.

Constructors

MkTrashEntryInfoNotFoundE (PathI TrashHome) (PathI TrashEntryFileName)

data TrashEntryInfoBadExtE

Unexpected file extension error.

Constructors

MkTrashEntryInfoBadExtE (PathI TrashEntryFileName) OsPath OsPath

Misc

newtype RenameDuplicateE

Could not rename file due to duplicate names.

Constructors

MkRenameDuplicateE (PathI TrashEntryPath)

data RestoreCollisionE

Collision with existing file when attempting a restore.

Constructors

MkRestoreCollisionE (PathI TrashEntryFileName) (PathI TrashEntryOriginalPath)

data RootE

Exception for deleting the root.

Constructors

MkRootE

data EmptyPathE

Exception for deleting an empty path.

Constructors

MkEmptyPathE

newtype DotsPathE

Exception for deleting the special dots paths.

Constructors

MkDotsPathE (PathI TrashEntryOriginalPath)

newtype FileNameEmptyE

Exception for deriving an empty file name.

Constructors

MkFileNameEmptyE (PathI TrashEntryOriginalPath)

data UniquePathNotPrefixE

Exception for when the original name is not a prefix of the derived unique name.

Constructors

MkUniquePathNotPrefixE (PathI TrashEntryFileName) (PathI TrashEntryFileName)

data InfoDecodeE

Exception for decoding.

Constructors

MkInfoDecodeE (PathI TrashEntryInfo) ByteString String

newtype EmptySearchResults

No search results.

Constructors

MkEmptySearchResults (UniqueSeq (PathI TrashEntryFileName))

newtype BackendDetectE

Unexpected backend error.

Constructors

MkBackendDetectE Backend