Charon.Runner.Args

Description

Provides CLI args functionality.

Documentation

getArgs :: MonadOptparse m => m Args

Retrieves CLI args.

data Args

CLI args.

Constructors

MkArgs

Fields

tomlConfigPath :: TomlConfigPath

Path to toml config.

backend :: Maybe Backend

Backend to use.

logLevel :: !(Maybe (Maybe LogLevel))

The file logging level. The double Maybe is so we distinguish between unspecified (Nothing) and explicitly disabled (Just Nothing).

logSizeMode :: Maybe FileSizeMode

Whether to warn/delete for large log files.

trashHome :: !(Maybe (PathI TrashHome))

Path to trash home.

command :: CommandP1

Command to run.

data TomlConfigPath

Toml path config.

Constructors

TomlNone
Do not use any Toml config.

TomlDefault
Attempts to read the Toml file at the default path.

TomlPath OsPath
Path to Toml file.