Charon.Runner.Toml

Description

Provides TOML configuration.

Documentation

data TomlConfig

Holds TOML configuration.

Constructors

MkTomlConfig

Fields

trashHome :: !(Maybe (PathI TrashHome))

Trash home.

backend :: Maybe Backend

Backend.

logLevel :: !(Maybe (Maybe LogLevel))

Log level. The double Maybe is so we distinguish between unspecified (Nothing) and explicitly disabled (Just Nothing).

logSizeMode :: Maybe FileSizeMode

Whether to warn/delete large log files.

mergeConfigs :: Args -> TomlConfig -> (TomlConfig, CommandP2)

Merges the args and toml config into a single toml config. If some field F is specified by both args and toml config, then args takes precedence.

defaultTomlConfig :: TomlConfig