Charon.Utils

Description

Provides internal utility functions

FAM combinators

throwIfEx :: (MonadIORef m, MonadThrow m) => IORef (Maybe SomeException) -> m ()

Throws the exception if it exists in the ref.

whenM :: Monad m => m Bool -> m () -> m ()

When for monadic bool.

Text

matchesWildcards :: Text -> Text -> Bool

matchesWildcards matchStr toMatch returns true if toMatch "matches" the matchStr, where unescaped asterisks in matchStr are interpreted as wildcards.

stripInfix :: Text -> Text -> Maybe (Text, Text)

stripInfix text needle strips the _first_ occurrence of needle from the text.

ByteString

breakEqBS :: ByteString -> (ByteString, ByteString)

Breaks a bytestring on the first '='. The '=' is removed from the second element.

Percent encoding

percentEncode :: ByteString -> ByteString

Percent encoded a bytestring.

percentDecode :: ByteString -> ByteString

Percent decodes a bytestring.

Optics

mergeAlt :: Alternative f => Lens' s (f a) -> Lens' t (f a) -> s -> t -> f a

Merge two fields using the Alternative instance.

merge :: (a -> a -> a) -> Lens' s a -> Lens' t a -> s -> t -> a

Merge two fields using the given function.

Bytes formatting

normalizedFormat :: Bytes B Natural -> Text

Normalizes and formats the bytes.

formatBytes :: (BaseFormatter (Unwrapped a) ~ FloatingFormatter, PrintfArg (Unwrapped a), Sized a, Unwrapper a) => a -> Text

Formats the bytes.

Logs

readLogLevel :: MonadFail m => Text -> m (Maybe LogLevel)

Reads the LogLevel.

logLevelStrings :: String

String description of possible log levels parsed by readLogLevel.

PathSize

getPathSize :: (HasCallStack, MonadAsync m, MonadCatch m, MonadLoggerNS m, MonadPathReader m, MonadPosixCompat m, MonadTerminal m) => OsPath -> m (Bytes B Natural)

getPathSizeIgnoreDirSize :: (HasCallStack, MonadAsync m, MonadCatch m, MonadLoggerNS m, MonadPathReader m, MonadPosixCompat m, MonadTerminal m) => OsPath -> m (Bytes B Natural)

getSymLinkSize :: (HasCallStack, MonadPosixCompat m, MonadThrow m) => OsPath -> m (Bytes B Natural)

Misc

filterSeqM :: forall m a. Monad m => (a -> m Bool) -> Seq a -> m (Seq a)

Filter a sequence monadically.

renderPretty :: Pretty a => a -> Text

Renders via pretty instance.

setRefIfJust :: MonadIORef m => IORef (Maybe a) -> Maybe a -> m ()

Sets the ioref if the maybe is non-empty.

noBuffering :: (HasCallStack, MonadHandleWriter m) => m ()

Sets NoBuffering.

getAllFiles :: (HasCallStack, MonadCatch m, MonadLoggerNS m, MonadPathReader m) => OsPath -> m [OsPath]

localTimeToMillis :: LocalTime -> Integer

getRandomTmpFile :: (HasCallStack, MonadLoggerNS m, MonadPathReader m, MonadThrow m, MonadTime m) => OsPath -> m OsPath